chore: remove unused file
