+++ script/buildRegistryItem.ts
@@ -1,5 +1,4 @@
-import { readdir, readFile, stat, mkdir, writeFile } from "fs/promises";
-import * as csstree from "css-tree";
+import * as fs from "fs/promises";
 import path from "path";
 
 interface RegistryItem {
@@ -18,43 +17,27 @@ interface RegistryItem {
   css?: Record<string, Record<string, Record<string, string>>>;
 }
 
-const targetFolderName = "../lib";
-const targetFolder = path.join(import.meta.dirname, targetFolderName);
-const OUTPUT_DIR = "public/r";
+const dirname = import.meta.dirname;
+const outputDirName = "../public/r";
+const inputDirName = "../lib";
 
-const readDirectory = async (currentPath) => {
-  const folderNames = await readdir(currentPath);
+const targetDir = path.join(dirname, inputDirName);
 
-  return folderNames;
+const EXCEPTIONS = ["stories"];
+
+// utils
+
+const filterDirectotries = async () => {
+  const allEntries = await fs.readdir(targetDir, { withFileTypes: true });
+  const filteredFolders = allEntries
+    .filter((entry) => entry.isDirectory() && !EXCEPTIONS.includes(entry.name))
+    .map((entry) => entry.name);
+
+  return filteredFolders;
 };
 
-const extractDependencies = (content: string): string[] => {
-  const dependencies = new Set<string>();
-  const importRegex =
-    /import(?:["'\s]*(?:[\w*{}\n\r\t, ]+)from\s*)?["'](.*?)["']/g;
-  let match;
-
-  while ((match = importRegex.exec(content)) !== null) {
-    const dep = match[1];
-    if (!dep.startsWith(".") && !dep.startsWith("node:")) {
-      if (dep.startsWith("@")) {
-        const parts = dep.split("/");
-        if (parts.length >= 2) {
-          dependencies.add(`${parts[0]}/${parts[1]}`); // @swc-react/button
-        } else {
-          dependencies.add(dep);
-        }
-      } else {
-        const parts = dep.split("/");
-        dependencies.add(parts[0]);
-      }
-    }
-  }
-  return Array.from(dependencies);
-};
-
-const getFileType = (filePath: string): string => {
-  const ext = path.extname(filePath).toLowerCase();
+const getFileType = (fileName: string): string => {
+  const ext = path.extname(fileName).toLowerCase();
   switch (ext) {
     case ".tsx":
     case ".jsx":
@@ -69,125 +52,90 @@ const getFileType = (filePath: string): string => {
   }
 };
 
-const main = async () => {
-  const folderNames = await readDirectory(targetFolder);
-  const stats = await Promise.all(
-    folderNames.map((item) => stat(path.join(targetFolder, item)))
-  );
-  const targetFileNames = folderNames.filter(
-    (item, index) => stats[index].isDirectory() && item !== "stories"
-  );
+const getDependencies = (content: string): string[] => {
+  const dependencies = new Set<string>();
 
-  await mkdir(OUTPUT_DIR, { recursive: true });
+  // 멀티라인 import 구문도 처리할 수 있는 정규식 (s 플래그 사용으로 . 이 줄바꿈도 매치)
+  const importRegex = /import\s+(?:[\s\S]*?\s+from\s+)?['"`]([^'"`]+)['"`]/g;
 
-  targetFileNames.forEach(async (item) => {
-    console.log(`Processing component: ${item}`);
-    if (item.includes("Provider")) return;
+  let match;
+  while ((match = importRegex.exec(content)) !== null) {
+    const importPath = match[1];
 
-    const currentPath = path.join(targetFolder, item);
-    const fileNames = await readdir(currentPath);
-    const allDependencies = new Set<string>();
+    // 상대 경로나 CSS 파일은 제외
+    if (
+      !importPath.startsWith(".") &&
+      !importPath.startsWith("/") &&
+      !importPath.endsWith(".css")
+    ) {
+      // @로 시작하는 scoped package나 일반 package name 추출
+      const packageName = importPath.startsWith("@")
+        ? importPath.split("/").slice(0, 2).join("/")
+        : importPath.split("/")[0];
 
-    const registryItem: RegistryItem = {
-      $schema: "https://ui.shadcn.com/schema/registry-item.json",
-      name: `serok ${item}`,
-      type: "registry:ui",
-      title: item,
-      description: `serok ${item}`,
-      dependencies: [],
-      files: [],
-    };
+      dependencies.add(packageName);
+    }
+  }
 
-    fileNames.forEach(async (file) => {
-      if (file.endsWith(".tsx")) {
-        const currentFilePath = path.join(currentPath, file);
-        const fileContent = await readFile(currentFilePath, "utf-8");
-        const fileDependencies = extractDependencies(fileContent);
-        fileDependencies.forEach((dep) => allDependencies.add(dep));
-        const fileType = getFileType(file);
-
-        registryItem.files.push({
-          path: currentFilePath,
-          type: fileType,
-          content: fileContent,
-        });
-      }
-
-      if (file.endsWith(".css")) {
-        const currentFilePath = path.join(currentPath, file);
-        const fileContent = await readFile(currentFilePath, "utf-8");
-        const ast = csstree.parse(fileContent);
-
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        const result: Record<string, any> = {};
-
-        csstree.walk(ast, {
-          visit: "Atrule",
-          enter(node) {
-            if (!node.prelude || node.block?.type !== "Block") return;
-
-            const layerName =
-              node.name === "layer"
-                ? `@layer ${csstree.generate(node.prelude)}`
-                : null;
-            if (!layerName) return;
-            if (!result[layerName]) result[layerName] = {};
-
-            for (const rule of node.block.children.toArray()) {
-              if (rule.type !== "Rule") continue;
-
-              const outerSelector = csstree.generate(rule.prelude);
-              if (!result[layerName][outerSelector])
-                result[layerName][outerSelector] = {};
-
-              for (const decl of rule.block.children.toArray()) {
-                if (decl.type === "Declaration") {
-                  result[layerName][outerSelector][decl.property] =
-                    csstree.generate(decl.value);
-                }
-              }
-
-              for (const nestedRule of rule.block.children.toArray()) {
-                if (nestedRule.type !== "Rule") continue;
-
-                const nestedSelector = csstree.generate(nestedRule.prelude);
-                result[layerName][outerSelector][nestedSelector] = {};
-
-                for (const decl of nestedRule.block.children.toArray()) {
-                  if (decl.type === "Declaration") {
-                    result[layerName][outerSelector][nestedSelector][
-                      decl.property
-                    ] = csstree.generate(decl.value);
-                  }
-                }
-              }
-            }
-          },
-        });
-
-        // console.log(JSON.stringify(result, null, 2));
-        registryItem.css = result;
-      }
-
-      if (registryItem.files.length > 0) {
-        if (allDependencies.size > 0) {
-          registryItem.dependencies = Array.from(allDependencies).sort();
-        }
-
-        const outputFilePath = path.join(
-          OUTPUT_DIR,
-          `${item.toLowerCase()}.json`
-        );
-
-        await writeFile(outputFilePath, JSON.stringify(registryItem, null, 2));
-        console.log(`Generated registry item for ${item}`);
-      } else {
-        console.warn(
-          `No .tsx or .css files found for component ${item}. Skipping.`
-        );
-      }
-    });
-  });
+  return Array.from(dependencies);
 };
 
-main();
+// build process
+(async function main() {
+  // 빌드 제외 폴더 및 파일을 제외
+  const filteredFolders = await filterDirectotries();
+
+  const targetPath = path.join(dirname, outputDirName);
+  await fs.mkdir(targetPath, { recursive: true });
+
+  // build start
+  await Promise.allSettled(
+    filteredFolders.map(async (folderName) => {
+      console.log(`build start: ${folderName}`);
+
+      try {
+        const sourcePath = path.join(targetDir, folderName);
+        const files = await fs.readdir(sourcePath);
+        const registry: RegistryItem = {
+          $schema: "https://ui.shadcn.com/schema/registry-item.json",
+          name: `serok ${folderName}`,
+          type: "registry:ui",
+          title: folderName,
+          description: `serok ${folderName}`,
+          files: [],
+          dependencies: [],
+        };
+
+        const allDependencies = new Set<string>();
+
+        await Promise.all(
+          files.map(async (file) => {
+            const filePath = path.join(sourcePath, file);
+            const content = await fs.readFile(filePath, "utf-8");
+            const fileType = getFileType(file);
+
+            if (fileType === "registry:ui") {
+              const fileDependencies = getDependencies(content);
+              fileDependencies.forEach((dep) => allDependencies.add(dep));
+            }
+
+            registry["files"].push({
+              type: fileType,
+              path: filePath,
+              content,
+            });
+          })
+        );
+
+        registry.dependencies = Array.from(allDependencies);
+
+        const outputFilePath = path.join(targetPath, `${folderName}.json`);
+        await fs.writeFile(outputFilePath, JSON.stringify(registry, null, 2));
+
+        console.log(`build success: ${folderName}`);
+      } catch (error) {
+        console.error(`build failed: ${folderName}`, error);
+      }
+    })
+  );
+})();
